Add files via upload
Hello all, I came across a sequence known as collatz sequence/conjecture. What happens here is that if you take a number such that if it is an odd number you multiply it itself 3 times and add 1 to it, and if it is even, you divide it by 2. the sequence will always end as 4,2,1. and then 4,2,1 becomes a loop(why? because n=(1*3)+1. and so on. Like always, I'm open to suggestions and feedbacks.